rename folder to ts
